+++ packages/code-du-travail-frontend/src/outils/DureePreavisLicenciement/steps/Result/Components/index.ts
@@ -1,0 +1,3 @@
+export { default as DisplayResult } from "./DisplayResult";
+export { default as DurationResult } from "./DurationResult";
+export { default as DisclaimerText } from "./DisclaimerText";